feat: 三端按钮镂空化统一 + 后端API契约治理 + 多端数据一致性
【按钮镂空化】（用户要求：蓝色边框+蓝色文字，全部镂空，业务色保留）
- PC端 4 处实心按钮改镂空：CrmDashboard、SmartAlertBell、StyleProcessTab、MaterialDatabase
- 小程序 5 文件按钮系统全部镂空化：app.wxss + scan-area/stage-detail/defect/admin
- H5端 6 文件按钮系统全部镂空化：form/login/scan/order/SampleInventory/CrmLogin
- H5 source-miniapp 同步 5 个 wxss（MD5 校验三端一致）
- 保留业务状态色（成功绿/危险红/警告黄），仅按钮改镂空

【后端API契约治理】前后端数据一致性
- 分页参数统一为 int page + int pageSize（8 个 Controller，禁止 String/Long/Integer 混用）
- 错误码语义化：16 处 Result.fail → badRequest(400)/notFound(404)
- 异常脱敏：33 处 catch 块 return Result.fail(e.getMessage()) → throw BusinessException
- Result.java 新增 notFound(404) 和 forbidden(403) 静态方法

【验证】
- mvn compile: BUILD SUCCESS
- npx tsc --noEmit: 0 errors
- python3 scripts/api_audit.py: 三端 0 不匹配（PC 549 + 小程序 263 + H5 208 调用全匹配）

【其他】
- .gitignore 追加 .design/ 和 *.design 规则，避免 IDE 工具产物误入库

+++ backend/src/main/java/com/fashion/supplychain/system/controller/OperationLogController.java
@@ -26,8 +26,8 @@ public class OperationLogController {
     @GetMapping("/list")
     @PreAuthorize("isAuthenticated()")
     public Result<?> getOperationLogList(
-            @RequestParam(defaultValue = "1") Long page,
-            @RequestParam(defaultValue = "10") Long pageSize,
+            @RequestParam(defaultValue = "1") int page,
+            @RequestParam(defaultValue = "10") int pageSize,
             @RequestParam(required = false) String module,
             @RequestParam(required = false) String operation,
             @RequestParam(required = false) String operatorName,
@@ -38,7 +38,7 @@ public class OperationLogController {
             @RequestParam(required = false) String endDate) {
 
         Page<OperationLog> logPage = operationLogService.getOperationLogPage(
-                page, pageSize, module, operation, operatorName, targetType, targetId, targetName, startDate, endDate
+                (long) page, (long) pageSize, module, operation, operatorName, targetType, targetId, targetName, startDate, endDate
         );
         return Result.success(logPage);
     }
@@ -52,10 +52,10 @@ public class OperationLogController {
     public Result<?> getOperationLogByTarget(
             @RequestParam String targetType,
             @RequestParam String targetId,
-            @RequestParam(defaultValue = "1") Long page,
-            @RequestParam(defaultValue = "20") Long pageSize) {
+            @RequestParam(defaultValue = "1") int page,
+            @RequestParam(defaultValue = "20") int pageSize) {
         Page<OperationLog> logPage = operationLogService.getOperationLogPage(
-                page, pageSize, null, null, null, targetType, targetId, null, null, null
+                (long) page, (long) pageSize, null, null, null, targetType, targetId, null, null, null
         );
         return Result.success(logPage);
     }
